fix: move config out
